+++ core/ajax/pyenv.ajax.php
@@ -37,6 +37,11 @@ try {
     }
   }
 
+  if (init('action') === 'ReinitPyenv') {
+    pyenv::reinit();
+    ajax::success();
+  }
+
   throw new Exception(__('Aucune méthode correspondante à', __FILE__) . ' : ' . init('action'));
   /*     * *********Catch exeption*************** */
 }